requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic[email]
pydantic-settings
PyJWT
pwdlib[argon2]
python-dotenv
passlib[argon2]
watchfiles
psycopg2-binary
httpx
